+++ src/main/java/com/github/alexisjehan/javanilla/lang/array/FloatArrays.java
@@ -29,6 +29,7 @@ import com.github.alexisjehan.javanilla.misc.quality.ToString;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
 
 /**
@@ -321,16 +322,32 @@ public final class FloatArrays {
 	}
 
 	/**
-	 * <p>Shuffle values in the given {@code float} array using the Fisher-Yates algorithm.</p>
+	 * <p>Shuffle values in the given {@code float} array following the Fisher-Yates algorithm.</p>
 	 * @param array the {@code float} array to shuffle
 	 * @throws NullPointerException if the {@code float} array is {@code null}
 	 * @see <a href="https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle">https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle</a>
 	 * @since 1.2.0
+	 * @deprecated for security purposes, use {@link #shuffle(float[], Random)} with
+	 *             {@link java.security.SecureRandom} instead
 	 */
+	@Deprecated(since = "1.6.0")
 	public static void shuffle(final float[] array) {
+		shuffle(array, ThreadLocalRandom.current());
+	}
+
+	/**
+	 * <p>Shuffle values in the given {@code float} array using the provided {@code Random} object following the
+	 * Fisher-Yates algorithm.</p>
+	 * @param array the {@code float} array to shuffle
+	 * @param random the {@code Random} object to use
+	 * @throws NullPointerException if the {@code float} array or the {@code Random} object is {@code null}
+	 * @see <a href="https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle">https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle</a>
+	 * @since 1.6.0
+	 */
+	public static void shuffle(final float[] array, final Random random) {
 		Ensure.notNull("array", array);
+		Ensure.notNull("random", random);
 		if (1 < array.length) {
-			final var random = ThreadLocalRandom.current();
 			for (var i = 0; i < array.length; ++i) {
 				swap(array, i, random.nextInt(i + 1));
 			}