+++ src/main/java/com/github/alexisjehan/javanilla/lang/array/ObjectArrays.java
@@ -31,6 +31,7 @@ import java.lang.reflect.Array;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import java.util.Random;
 import java.util.WeakHashMap;
 import java.util.concurrent.ThreadLocalRandom;
 
@@ -345,16 +346,32 @@ public final class ObjectArrays {
 	}
 
 	/**
-	 * <p>Shuffle values in the given {@link Object} array using the Fisher-Yates algorithm.</p>
+	 * <p>Shuffle values in the given {@link Object} array following the Fisher-Yates algorithm.</p>
 	 * @param array the {@link Object} array to shuffle
 	 * @throws NullPointerException if the {@link Object} array is {@code null}
 	 * @see <a href="https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle">https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle</a>
 	 * @since 1.2.0
+	 * @deprecated for security purposes, use {@link #shuffle(Object[], Random)} with
+	 *             {@link java.security.SecureRandom} instead
 	 */
+	@Deprecated(since = "1.6.0")
 	public static void shuffle(final Object[] array) {
+		shuffle(array, ThreadLocalRandom.current());
+	}
+
+	/**
+	 * <p>Shuffle values in the given {@link Object} array using the provided {@code Random} object following the
+	 * Fisher-Yates algorithm.</p>
+	 * @param array the {@link Object} array to shuffle
+	 * @param random the {@code Random} object to use
+	 * @throws NullPointerException if the {@link Object} array or the {@code Random} object is {@code null}
+	 * @see <a href="https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle">https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle</a>
+	 * @since 1.6.0
+	 */
+	public static void shuffle(final Object[] array, final Random random) {
 		Ensure.notNull("array", array);
+		Ensure.notNull("random", random);
 		if (1 < array.length) {
-			final var random = ThreadLocalRandom.current();
 			for (var i = 0; i < array.length; ++i) {
 				swap(array, i, random.nextInt(i + 1));
 			}